Update IFNode seq number even for Delete operation
Lets say if Node A, Node B and link L1 exists between them with
each entity's seq number being 1. After Xmpp channel flap we generate
a new seq number for the subsequent config. After XMPP flap, if an
update is receive for Node B, its seq number would be 2 with A and L1
being at seq 1. If Node A's delete is received before L1 delete,
L1  would be pushed to Defer list. While pushing to defer list,
the seq number of defer entry A->B is taken as 2 (as B's seq is 2)
and B->A entry is taken as 1 (as A's seq is 1). This is leading to
dissimilar seq number for A->B and B->A. When subsequntly add of Node A
is received, while evaluating defer list, B->A is not considered as its
seq is old seq leading to deletion of defer enty A->B but not B->A. This
ends up in a single defer entry being present which is invalid.

As a fix, when Node A's delete is received, the seq number is of A is
updated before pushing its links to Defer list. While adding defer list,
the seq number of Defer entry is taken from IFMapLink rather from
end nodes. While accepting an update to a node in DB,  and assert is
added to ensure that update's seq number is same or higher than existing
node. While evaluating Defer list, it is ensured that both direction
entry's are removed rather a single.

Change-Id: Ic9ff72a51f0aa683266792a147e6b0a7a8be7699
closes-bug: #1465190