Clean up all routing instances when a service chain is deleted
If a service chain is not created completely (perhaps due to vm or vmi
not having been received yet by schema transformer), and it is deleted,
schema transformer does not clean the routing instances created earlier.
With this change, we track if a service chain is created partially and
clean it up at the time of deletion.

In mainline, we will add checks to not even create routing instances if
all data to create the chain is not available to schema transformer.

Change-Id: I7078dc69fde8149d3962dafe746dc51b63277b46
Closes-Bug: 1461058