Instance manager task queue size
To insert the Instnace task into task queue, hash of service instance
uuid is used. To identify the numer of task queues avaialble,
.capacity() routine is used on vector, whose value mgiht be more than
the size of the vector. If task is inserted into a vector position
that is more than the size of vector, exception can occur. To avoid
this, .size() routine is invoked.

Change-Id: I8035ebfcdc3b146e24976cd95aad83bb9f95283c
closes-bug: #11426