Check for virtual-network-network-id when VN is deleted
Earlier, vn id was stored as part of virtual-network-properties, but was
later changed to store as standalone property. When the networks is deleted,
we need to check at both places to get the network id and free it.

Change-Id: Ib70ef6c6aa638de684f290933579798ef94f999c
Closes-Bug: 1467638
(cherry picked from commit 2296246a3aa5a30927bfe39f702a03131df928af)