Force peer validity check for BGP peer.
Extend the Peer class to specify if a validity check for peer
is needed whenver it is enqueued as a key paramater for DB request.
Currently BGP peer derived class falls into this category.
Any agent route data enqueued with BGP peer in key will have to
implement validity check.

Change-Id: I47299571832c168c755da4a419fe95812dab5ba9
Closes-bug: 1466739