Force peer validity check for BGP peer.
Extend the Peer class to specify if a validity check for peer
is needed whenver it is enqueued as a key paramater for DB request.
Currently BGP peer derived class falls into this category.
Any agent route data enqueued with BGP peer in key will have to
implement validity check.

Change-Id: Ic69ad46ea2e9cf8e63eafdf67f37d3caafaf1b89
Closes-bug: 1466739