Do not send interface stats UVE when there are no change in statistics
VMI stats (diff stats since previous send) are not sent if the diff stats are 0.
Tags annotation is removed for bandwidth and now the bandwidth is sent from agent only if it has changed since last send.
Also remove "virtual_network" tag for VMI table as it is not being used.

Change-Id: I46f95405e577d005326e3841d7f6bcc6827c999a
Closes-Bug: #1469632