PathPreference state not cleaned.
Problem:
When path preference is deleted because of route table delete, walk is started
to delete all the states. It can so happen that before walk done is called,
notification is received for one of the table entry. This will result in state
addition again.

Solution:
If delete has been issued ignore any notifications on same.

Change-Id: I83c18d52d7e0a42eb47d92f429dce9fc8cf9e2fa
Closes-bug: 1470736