PathPreference state not cleaned.
Problem:
When path preference is deleted because of route table delete, walk is started
to delete all the states. It can so happen that before walk done is called,
notification is received for one of the table entry. This will result in state
addition again.

Solution:
If delete has been issued ignore any notifications on same.

Change-Id: Id8b991ee295ba0010ad63feb11558f20965848ed
Closes-bug: 1470736