PathPreference state not cleaned.
Problem:
When path preference is deleted because of route table delete, walk is started
to delete all the states. It can so happen that before walk done is called,
notification is received for one of the table entry. This will result in state
addition again.

Solution:
If delete has been issued ignore any notifications on same.

Change-Id: I6b60bb6ad5acfe701a905e356bb4858c6c50a627
Closes-bug: 1470736