Explicitely drop Cassandra keyspaces
In the case of schema or svc_monitor the connection to
VncCassandraClient should not drop the Vnc keyspace when reset_config is
used. This connection is only used for reading.

Only keyspaces owned by schema and svc_monitor should be droped when
using reset_config.

The problem affects contrail-installer since all services are run with
reset_config=True. In this case if the api server is run _before_ schema
or svc_monitor its keyspace will be droped.

The VncCassandraClient takes now a list of keyspaces names to reset in
the reset_config parameter instead of a boolean.

Change-Id: I96bddd49a501cc2adfab3faf1dd79ea07ad4fd0c
Closes-Bug: #1472197