Add config-tracker database inside each client.
Today, when the client is going away, we do a graph-walk visiting each node
starting at the client's VR and reset its state bits. This walk will not be
able to reach all the relevant nodes for this client in all the cases.
EG scenario:
1. Exporter processes link-delete that partitions the graph. Asks GW to process
it. The graph is partitioned such that a portion of the graph is not accessible
to the client anymore.
2. Client delete (NOT_READY) happens causing the client to cleanup and
disappear.
3. GW runs but wont be able to clear the deleted clients bit since its gone.

Without #2, #3 would cleanup all relevant bits. But, with #2, it wont happen
since #2 will only cleanup nodes still reachable via its VR and #3 wont be able
to cleanup the unreachable nodes since the client object would be gone by the
time #3 happens.

Fix is for #2 to cleanup the state's of all nodes relevant for this client.
Instead of a brute-force walk of all nodes in all tables, we will now maintain
a unordered-set inside each client of all nodes it has touched. While
processing the client delete, we will go through this unordered-list and
cleanup all relevant nodes.

Closes-Bug: 1472807

Change-Id: I0890106f46ccc54d650071a2421698f5642d817a