Introspect changes to display config-tracker information for each client.
1. Add backpointers to node/link from the state via IFMapObjectPtr so that we
can access the node/link via the state. To do this, we need to change the
constructors for IFMapNodeState and IFMapLinkState.
2. Add information to the per-client node/link introspect frontend to show the
tracker info. Also, add a boolean to show the tracker information even if the
interest bit is not set for the client in the node/link i.e. stale entries left
behind because of some bug.
3. If the user does not give any client index as input, make sure we return
nothing. Today we return the information for client with index zero. This
happens because strtoul() returns zero and sets endptr[0] to '\0' when the
string is empty which is undesirable. The intention always was to get the index
from the client. We are enforcing it now i.e. if the index is not entered, we
return empty output.

Change-Id: Iebb19ce108f30a35cc594e3fbd0c57229341d05a
Closes-Bug: #1472807