Introspect changes to display config-tracker information for each client.
1. Add information to the per-client node/link introspect frontend to show
the tracker info.
2. If the user does not give any client index as input, make sure we return
nothing. Today we return the information for client with index zero. This
happens because strtoul() returns zero and sets endptr[0] to '\0' when the
string is empty which is undesirable. The intention always was to get the index
from the client. We are enforcing it now i.e. if the index is not entered, we
return empty output.

Closes-Bug: #1472807

Change-Id: Ifc683c6e576f02aaabe7b02da8a7643ebbb7c419