Support pagination of output for route summary introspect
Limit the maximum number of entries displayed on a single page. A
next_batch link is generated if there are more entries to display.

Also limit maximum number of entries examined in one invocation of
the callback routine.  This comes into play when there is a search
string specified and many entries don't match it.  A partial page
is saved in user-allocated data and the next invocation of callback
appends to it.  This is repeated till there's a full page or there
are no more entries in the table.

Following changes are implemented:

- Move code from bgp_sandesh.cc to bgp_show_route_summary.cc
- Implement iteration limit to avoid hogging CPU from introspect
- Add unit tests to cover combinations of page and iteration limits
- Remove older unit tests which were limited in scope

Change-Id: I07657e8b4303a8c3e4c6ffe307a751f36e55da9b
Closes-Bug: 1477669