Generate mac from instance ip for service VMs
Service instance VM, netns do not request mac-addresses for its ports.
In cases of active-standby, active-active instances there is an issue
which can cause the standby mac-address to be picked up. This change
ensures that the same mac-address is used for all interfaces that share
the same instance ip address.

Change-Id: I409eae920216841656c18e947b7f8d49ef14ab04
Closes-Bug: #1460767