Add more functionality for ifmap stress test.
1. Add support for real xmpp clients using IFMapXmppClientMock. The clients are
created during setup.
2. Add support for events like VR/VM node creation/deletion, VR subscribe, VM
subscribe/unsubscribe, XMPP connection creation via IFMapXmppClientMock. Add
callback support to call these routines when the corresponding event is chosen.
3. Simulate XMPP state machine by doing the following:
a. Keep track of clients that have created the XMPP connection. Use this group
to pick clients for VR-subscribe and keep track of clients with successful
VR-subscribes.
b. For each client, maintain a set of vm-ids that they will host. For
VM-subscribes, pick a client from the successful VR-subscribe group. For that
client, pick a VM that has not subscribed and keep track of subscribed VMs.
For VM-unsubscribes, pick a client from the successful VR-subscribe group. For
that client, pick a VM that has already subscribed.
4. Add counters and log support to keep track of the sequence of events that
occurred in the test. The logs are stored in a circular buffer.
5. Add number of VMs per client as a command line parameter.

Partial-Bug: #1469909

Change-Id: I72cb8c09582cb41949f317bf14868fc7708a169c