Handle 'locate' failures in service-monitor
Encountered a case where there were inconsistent index <-> objects in config db
(i.e. fq-name-to-id said object was present but in fact object was not present).
In this case service monitor failed in locate of this object (could be VMI/VN/SI)
and service would forever be unavailable.
Added exception handling to log and skip the missing objects.

Change-Id: Idd3400742f2c3e207d4c149a178524780fa0b003
Closes-Bug: #1480402