Make class XmppVnswMockPeer generic
1. Move class XmppVnswMockPeer from ifmap_xmpp_test.cc to a new file
ifmap_xmpp_client_mock.cc so that it can be used generically.
2. Replace all #defines with static const's or local variables in the tests
and change all tests to use them.
3. Add username to the output file name created by each test so that there is
no filename/owner clash if someone else runs the test. If a file with that name
already exists since it was created by a different user, file open fails and
the reference file is not created.
4. Completely revamp the constructor of IFMapXmppClientMock to take parameters
instead of using globally defined #defines. Also, pass proper default
parameters for them instead of passing empty strings and then assigning inside
the constructor body when the strings are empty.
5. Remove all superfluous headers from the test file.

Closes-Bug: #1479043

Change-Id: I1fdf288232fac133e3f306863b6aa2fc50016929