Support pagination of output for bgp neighbor introspect
Limit the maximum number of entries displayed on a single page for
both regular and summary requests.  A next_batch link is generated
if there are more entries to be displayed.

Also limit maximum number of entries examined in one invocation of
the callback routine.  This comes into play when there is a search
string specified and many entries don't match it.  A partial page
is saved in user-allocated data and the next invocation of callback
appends to it.  This is repeated till there's a full page or there
are no more entries in the table.

Following changes are implemented:

- Move code from bgp_sandesh.cc to bgp_show_neighbor.cc
- Use class template BgpShowHandler to avoid code duplication
- Implement iteration limit to avoid hogging CPU from introspect
- Add unit tests to cover combinations of page and iteration limits
- Remove older unit tests which were limited in scope
- Sprinkle const as required

Change-Id: I4ef1848fb315ebbf4738e21f313cf60c255b7912
Closes-Bug: 1479427