* If a ECMP flow gets processed as layer-2 flow mark it as short flow
If source route is composite, then the flow has to be layer3
procesess so that proper load distribution happens, this is
forced but proxying with vrouter mac for ECMP destinations
and if ECMP source originates the packet then also vrouter
would proxy with vrrp mac so that packet gets processed
as layer3 flow. Remove the assert in agent and mark the
flow as short flow.
Closes-bug:#1472796

Change-Id: I42df08e883d69779fc0a45f44098d3b86035a97c