Check size of MAC vector before reading the element
While clearing the neutron router, Agent some times receives NULL object
without any properties. As this is not a delete operation
ServiceInstance attempts to calcualte the properties of the instance by
traversing the node grpah. When VMI is seen, it attempts to read the mac
address which as such is not preset leading to crash.

As a fix, the mac address is read only if the Mac vector has some
element in it.

Change-Id: Idd2afde67f56728408e5c553d772f30d00a246af
closes-bug : #1474272