Enable tcp level timeout to detect when the remote end has crashed or network connectivity is lost so we could trigger rebuild of the multicast tree.
Setting idle-timeout to 15 seconds after which we trigger keepalive
probes every 3 secs for 5 times, to detect inactivity when the
socket buffer is empty.

We also need to set tcp-user timeout to detect inactivity when
the socket transmit buffer is full.

In both cases, the timeout is 30 secs.

Current application level hold-timer will detect peer DOWN after 90secs,
the time taken to to rebuild the multicast tree.

Change-Id: Idd821aba8fac6a7f31bc7874568e8584a9292994
Closes-bug:1479198