Set TCP_USER_TIMEOUT on connections to ifmap-server.
This is required for cases where the control-node's send buffer is non-empty
and the other end becomes unreachable. Without this option, we will wait for
20 mins before giving up on the connection. During this time, the config could
change but the control-node will not get any updates since the ifmap-server is
unreachable.

Using this option and then reconnecting to a different reachable ifmap-server
solves the problem.

Note, this is not required for cases where the other end crashes since we will
get a reset in that case. This is required only for reachability problems.

Change-Id: Ie0d5c92a76d8913177e137bd0eacbc6a033119ec
Closes-Bug: #1479521
(cherry picked from commit 064379c3447c44a54f77575a93cbfd380a7a73f9)