Removing the config listener
After IFmap dependency manager is introduced for all the objects in Agent,
the config listener does not play any role other than invoking node observers
and link observers, which is taken care in dependency manager itself.

As part of the same, Uuid change of node is also detected and handled.

Couple of test cases are moved out of flaky tests.

Change-Id: Id2b342f01400e213f3269246c80459f257d10b6d
closes-bug: #1480124