Avoid trace when listing SG with non-existing tenant_id
Currently when listing SGs with tenant_id that doesn't exist
we get a stracktrace. The behaviour of neutron is to create
the SG we the tenant_id given which doesn't seem acceptable
for contrail. This patch remove the traceback and return nothing
when listing SGs with a non existing tenant. This also removes an
API call.

Change-Id: I5521c7630036ed28f6c87ed542204800086e50bc
Partial-Bug: #1480901