Avoid trace when listing SG with non-existing tenant_id
Currently when listing SGs with tenant_id that doesn't exist
we get a stacktrace. The behaviour of neutron/ML2 is to create
a default SG for the tenant_id given even if the tenant
doesn't exist. For contrail is seems to be better to return
an empty answer is this case.

This patch remove the traceback and return nothing
when listing SGs with a non existing tenant id.

Change-Id: I5521c7630036ed28f6c87ed542204800086e50bc
Partial-Bug: #1480901