Use the next-hop specified in host route DHCP option.
The nexthop provided in host route DHCP option is not being used, as we always
used the subnet's GW address. With L2 networks, contrail may not have GW
configured and external GW may have to be used. Changing this behaviour such
that the specified GW is sent and when the GW is 0.0.0.0, the subnet's GW
is used.

closes-bug: 1483105
(cherry picked from commit e924ed85f70219fe045136c0ca6599d1972255d7)

Conflicts:
	src/vnsw/agent/oper/vn.cc

Change-Id: I629ca9b7daa4363ffc9c101cf29210e1204f51e1