Use the next-hop specified in host route DHCP option.
The nexthop provided in host route DHCP option is not being used, as we always
used the subnet's GW address. With L2 networks, contrail may not have GW
configured and external GW may have to be used. Changing this behaviour such
that the specified GW is sent and when the GW is 0.0.0.0, the subnet's GW
is used.

Change-Id: Idf603d6beabc61558a227fc1f1d5046feb26502a
closes-bug: 1483105