Implement introspect for evpn information
Limit the maximum number of entries displayed on a single page for
both regular and summary requests.  A next_batch link is generated
if there are more entries to be displayed.

Also limit maximum number of entries examined in one invocation of
the callback routine.  This comes into play when there is a search
string specified and many entries don't match it.  A partial page
is saved in user-allocated data and the next invocation of callback
appends to it.  This is repeated till there's a full page or there
are no more entries in the table.

Following changes are implemented:

- Count number of unicast and inclusive multicast routes in EvpnTable
- Use class template BgpShowHandler to avoid code duplication
- Implement iteration limit to avoid hogging CPU from introspect
- Reuse show routing instance typed tests for evpn table introspect

Pending changes:

- Add more information to response for non-summary request
- Add unit tests to verify the contents of said response

Change-Id: I5f10a33e37b5b356d575400dce12303e10394540
Partial-Bug: 1479430