Prevent instance/table from getting deleted while iterating
The call to resume deletion of paused instances and tables happens
in the context of the main thread.  That could cause an instance or
table to get deleted via the Bgp::Config Task while the main thread
is iterating over them.

Fix by locking the scheduler when iterating over the instances and
tables.

Change-Id: I056119b72b8136764845c72334d58d8510961918
Closes-Bug: 1477669