Support pagination of output for multicast manager introspect
Limit the maximum number of entries displayed on a single page.
A next_batch link is generated if there are more entries to be
displayed.

Also limit maximum number of entries examined in one invocation of
the callback routine.  This comes into play when there is a search
string specified and many entries don't match it.  A partial page
is saved in user-allocated data and the next invocation of callback
appends to it.  This is repeated till there's a full page or there
are no more entries in the table.

Following changes are implemented:

- Move code from bgp_sandesh.cc to bgp_show_multicast_manager.cc
- Use class template BgpShowHandler to avoid code duplication
- Implement iteration limit to avoid hogging CPU from introspect
- Reuse type parameterized tests to also cover multicast manager
- Remove older unit test which was limited in scope
- Sprinkle const as required

Change-Id: I190cfbaaa58ac52ce63fb6bd916c040ab100e742
Closes-Bug: 1479428