Wrong port id in floatingip-show of floatingip assigned to virtual-ip port
In case of floating ip on the Virtual-ip, svc-monitor will link floating ip to
"right" interface of service VMs launched by ha-proxy service instance.

Ignore such VMI while walking the vmi_ref from floatingip object.
This is done based on the name of the interface.
SVC mon creats the serviceinstance interface with "__right__" in it.

Change-Id: Icb8cc874da4d18c4def631566fc7e2257f2f5fe4
Closes-bug: #1482277