TOR agent mcast route updated without active connection.
Tor-agent should export route only when OVS peer has added a path.
By default a local peer path is added irrespective of mode in mcast.
This was to subscribe in TSN mode. However this should not be used for exporting
in Tor agent mode.

Solution is to check for tor agent mode and skip subscription.

Change-Id: Ibc655d90973e5afd5e47826c74c6ff0696c60625
Closes-bug: 1487567