Null check for VRF state.
Problem:
Say a walk for notification of routes(NOTIFYALL) was going on for a peer.
It had enqueued the route walk because of same. However before dequeue happens,
vrf delete is issued. Notification of same will remove state from VRF for the
peer listener id(one which had started the walk above). After this dequeue
happens and route walk is started. RouteNotify then queries for vrf state(which
                                                                          is
                                                                          deleted)
and then accesses the same without null check.

Soltion:
Obvious

Change-Id: Ia2516c0928f62644ebe5a4ccabc3498395f83885
Closes-bug: 1487647