Removing the config listener
After IFmap dependency manager is introduced for all the objects in Agent,
the config listener does not play any role other than invoking node observers
and link observers, which is taken care in dependency manager itself.

As part of the same, Uuid change of node is also detected and handled.

Couple of test cases are moved out of flaky tests.

Change-Id: I82e646b97004ad96f799ad10cfbcca5a2823b3f1
closes-bug: #1480124