Mac was not stitched for non-local routes from CN.
Problem:
VN and VRF object gets added however link between them comes later.
By the time link comes, agent has subscribed for this VRF and receives routes.
This results in evaluation of mac binding in ksync. Since link is not present
yet VN in VRF oper object is NULL. So mac binding is skipped (For binding, vn is
                                                              needed in VRF to
                                                              extract forwarding
                                                              mode).
Now link comes and VN is populated in VRF but mac binding is not updated because
there is no reason seen to resync route.
This results in the issue mentioned in bug.

Solution:
When VN gets added in VRF oper object, resync routes.

Change-Id: Iabc10b4f6c09862fdad449afc2622971961ae634
Closes-bug: 1488548