Segregate XMPP message traces for config and route.
Insert a hook in XMPP to use custom message tracer.
In case no custom message tracer is present all messages default to XmppTrace.
With this commit a new trace handler is added for pulling in config and route RX
messages from control-node. They are pushed in respective traces meant for them.
This helps in scaled out setup.

Conflicts:
	src/vnsw/agent/controller/controller_init.cc

Change-Id: Icef08a5dcf81f4da3a1f688d7b3bb3c8d4ab96c5
Closes-bug: 1490891