Delete state only after processing a DELETE IFMapUpdate.
There can be cases when while processing an UPDATE-IFMapUpdate dequeue event,
we find that the state is ready to be deleted. We want to wait until we have
processed the 'DELETE' IFMapUpdate event to finally delete the state.

EG: A node add, link add with that node, link delete, node delete, exporter
processes node delete before link delete i.e. dependency is still set and
exporter does not enqueue a DELETE-UPDATE-IFMapUpdate. Now, if we dequeue
UPDATE-IFMapUpdate, the state is ready to be deleted. But, we want to wait
until exporter processes the link-delete and triggers a node-change via
MaybeNotifyOnLinkDelete() to delete the state.

Change-Id: I42ebbd5a9cff824c9732c52a773b6bc9ee36dc1a
Closes-Bug: #1484784