Handle deletion of vmi correctly in schema transformer
Currently, schema transformer only deletes VMi from its internal database when
link to VM goes away. This causes problems for VMIs which are not attached to
any VMI (e.g. VMI created for TOR). As a result, when another VMI with same name
is created, schema transformer does not link it to the default RI. With this
change, if the link to its parent goes away, we will delete the VMI locally.

Change-Id: Ic17a31f67a3ca7ee6d7c136bf1d348224f561646
Closes-Bug: 1477371