Agent crash seen for db entry which is deleted.
Problem:
In the mentioned bug a VRF DB entry was seen which was present in DB table with
sll references removed (This happens when VRF is marked for delete from config
however delete is deferred till all references go off)
Now this entry is accessed during a walk for deleted peer. This walk for vrf
tries to take a reference for enqueing route walk in that vrf. For the above
entry there was no state present but this reference count results in state
add(for reference). This is caught as assert in set_state (Entry scheduled for
                                                           delete is adding
                                                           state/reference when
                                                           all has been
                                                           removed).

Solution:
Avoid scheduling route walk(delpeer) for VRF where no state is present for bgp
peer. In above case there was no state for peer, so route walks will not be
scheduled and in turn no reference will be tried.

Change-Id: Id378271ada77cdd2cd837b312140c41b1280a937
Closes-bug: 1495824
(cherry picked from commit dc554e7c024f23548e60a254aac85a58a08e9e9a)