Do not check for presence of ports in OS in DPDK mode
In case of Agent restarts, agent reads all ports that were active in agent before it died from a specific directory. The read port information is validated
by checking for their presence in OS. If the are not present in OS, then they are not added. This check needs to be skipped in DPDK mode.
Closes-Bug: #1497012

Change-Id: Ib2f34a853e11f62759d297b0413618cd4eda430c