Fix the incorrect logic between RecomputeInterest() and CleanupInterest()
When the graph-walker processes a link-delete for a group of clients, for each
client, it does a combination of RecomputeInterest() and CleanupInterest() via
a graph-walk. But, other events can occur between the time the link-delete is
queued for the graph-walker and the graph-walker actually processes it.
Consider the following sequence:
1) exporter queues link-delete to the graph-walker for a client-group.
2) a completely different link-add processed by the
server-table, new link added to the graph, queued for the listeners.
3) graph-walker processes #1 via a graph-walk. done_set/rm_mask_ will have
client C1 when we are done.
4) client-delete for one of the clients, C1 in the client-group.

When we do a graph-walk in #3, there are 4 cases that can happen:
a. A node was reachable before #1 and is reachable during #3.
b. A node was reachable before #1 and is un-reachable during #3.
c. A node was un-reachable before #1 and is reachable during #3 because of #2.
d. A node was un-reachable before #1 and is un-reachable during #3.

Our current code does not handle 'c'. RecomputeInterest() marks the node as
reachable when we do #3. But, CleanupInterest() only examines nodes that were
reachable before #1. So, the nodes of type 'c' will continue to have 'nmask'
set with the bit for client dying in #4. Nobody will clean this bit from nmask.
Eventually, this bit can get propogated to other nodes and links although the
client is dead.

Fixes:
1. Add code to keep track of nodes of type 'c' via new_reachable_nodes_tracker_.
2. Use new_reachable_nodes_tracker_ to handle nodes of type 'c' in
LinkDeleteWalkBatchEnd()/NewReachableNodesCleanupInterest(). Use
OldReachableNodesCleanupInterest() to process 'a' and 'b' as we did before.
3. Remove rm_mask_ from the graph-walker since its no longer required.
Instead, pass the client index to CleanupInterest() to be used as rm_mask. We
needed rm_mask_ earlier since CleanupInterest() was called for all clients in
one go. But, since we use the config-tracker now, we work with one client at a
time.
4. Fixing a few things in LinkDeleteWalk(). For each bit in
link_delete_clients_, the client must exist and so we can assert instead of
checking if it exists. We dont use rm_mask_ anymore and instead we just pass
done_set to LinkDeleteWalkBatchEnd().

Change-Id: Ibb181a79d57712e066f767f3ad4db6a382a04590
Closes-Bug: #1495594